+++ conda_lock/src_parser/conda_common.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from .._vendor.conda.models.channel import Channel
 from .._vendor.conda.models.match_spec import MatchSpec
-from ..src_parser import VersionedDependency
+from ..src_parser.models import VersionedDependency
 
 
 def conda_spec_to_versioned_dep(spec: str, category: str) -> VersionedDependency:
