+++ conda_lock/src_parser/meta_yaml.py
@@ -6,7 +6,7 @@ import jinja2
 import yaml
 
 from conda_lock.common import get_in
-from conda_lock.src_parser import Dependency, LockSpecification, aggregate_lock_specs
+from conda_lock.src_parser.models import Dependency, LockSpecification
 from conda_lock.src_parser.selectors import filter_platform_selectors
 
 
@@ -92,6 +92,8 @@ def parse_meta_yaml_file(
     * This does not support multi-output files and will ignore all lines with
       selectors other than platform.
     """
+    from conda_lock.src_parser import aggregate_lock_specs
+
     # parse with selectors for each target platform
     spec = aggregate_lock_specs(
         [
